[PF-1930] Support cloning controlled bucket to referenced bucket
[PF-1930] Support cloning controlled bucket to referenced bucket

[PF-1930] Support cloning controlled bucket to referenced bucket

[PF-1930] Fix dismal failure with CreateStorageTransferServiceJobStep undo

fix test

[PF-1930] Support cloning controlled bucket to referenced bucket

fix unit tests

[PF-1930] Fix dismal failure with CreateStorageTransferServiceJobStep undo

fix COPY_DEFINITION -> COPY_REFERENCED

implement undo

have dest bucket use source bucket name

fix connected tests

fix comment

rebase fixes

update test

call workspaceService.deleteWorkspace

typo

ControlledResourceServiceTest tests are done

remove subflight

revert CopyGcsBucketDefinitionStep

fix rebase

unit test for bad request

fix rebase

+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/bucket/CloneControlledGcsBucketResourceFlight.java
@@ -17,6 +17,7 @@ import bio.terra.workspace.service.resource.controlled.flight.update.RetrieveCon
 import bio.terra.workspace.service.resource.controlled.model.ControlledResource;
 import bio.terra.workspace.service.resource.model.CloningInstructions;
 import bio.terra.workspace.service.resource.model.WsmResourceType;
+import bio.terra.workspace.service.resource.referenced.flight.create.CreateReferenceMetadataStep;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
 import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ResourceKeys;
@@ -29,12 +30,13 @@ import java.util.UUID;
 // 2. Gather controlled resource metadata for source object
 // 3. Gather creation parameters from existing object
 // 4. Launch sub-flight to create appropriate resource
-// Steps 5-9 are for resource clone only
-// 5. Set bucket roles for cloning service account
-// 6. Create Storage Transfer Service transfer job
-// 7. Listen for running operation in transfer job
-// 8. Delete the storage transfer job
-// 9. Clear bucket roles
+// 5. If cloning to referenced resource, do the clone and finish flight.
+// Steps 6-10 are for cloning to controlled resource only
+// 6. Set bucket roles for cloning service account
+// 7. Create Storage Transfer Service transfer job
+// 8. Listen for running operation in transfer job
+// 9. Delete the storage transfer job
+// 10. Clear bucket roles
 public class CloneControlledGcsBucketResourceFlight extends Flight {
 
   public CloneControlledGcsBucketResourceFlight(
@@ -97,29 +99,49 @@ public class CloneControlledGcsBucketResourceFlight extends Flight {
               flightBeanBag.getGcpCloudContextService(),
               RetrievalMode.CREATION_PARAMETERS),
           cloudRetry);
-      addStep(
-          new CopyGcsBucketDefinitionStep(
-              userRequest,
-              sourceBucket,
-              flightBeanBag.getControlledResourceService(),
-              resolvedCloningInstructions));
 
-      if (CloningInstructions.COPY_RESOURCE == resolvedCloningInstructions) {
+      if (CloningInstructions.COPY_REFERENCE == resolvedCloningInstructions) {
+        // Destination bucket is referenced resource
         addStep(
-            new SetBucketRolesStep(
+            new SetReferencedDestinationGcsBucketInWorkingMapStep(
+                userRequest,
                 sourceBucket,
-                flightBeanBag.getGcpCloudContextService(),
-                flightBeanBag.getBucketCloneRolesService(),
-                flightBeanBag.getStoragetransfer()),
-            cloudRetry);
+                flightBeanBag.getReferencedResourceService(),
+                resolvedCloningInstructions),
+            RetryRules.shortExponential());
         addStep(
-            new CreateStorageTransferServiceJobStep(flightBeanBag.getStoragetransfer()),
-            cloudRetry);
-        addStep(new CompleteTransferOperationStep(flightBeanBag.getStoragetransfer()), cloudRetry);
+            new CreateReferenceMetadataStep(flightBeanBag.getResourceDao()),
+            RetryRules.shortDatabase());
+        addStep(new SetReferencedDestinationGcsBucketResponseStep(), RetryRules.shortExponential());
+        return;
+      } else {
+        // Destination bucket is controlled resource
         addStep(
-            new DeleteStorageTransferServiceJobStep(flightBeanBag.getStoragetransfer()),
-            cloudRetry);
-        addStep(new RemoveBucketRolesStep(flightBeanBag.getBucketCloneRolesService()), cloudRetry);
+            new CopyGcsBucketDefinitionStep(
+                userRequest,
+                sourceBucket,
+                flightBeanBag.getControlledResourceService(),
+                resolvedCloningInstructions));
+
+        if (CloningInstructions.COPY_RESOURCE == resolvedCloningInstructions) {
+          addStep(
+              new SetBucketRolesStep(
+                  sourceBucket,
+                  flightBeanBag.getGcpCloudContextService(),
+                  flightBeanBag.getBucketCloneRolesService(),
+                  flightBeanBag.getStoragetransfer()),
+              cloudRetry);
+          addStep(
+              new CreateStorageTransferServiceJobStep(flightBeanBag.getStoragetransfer()),
+              cloudRetry);
+          addStep(
+              new CompleteTransferOperationStep(flightBeanBag.getStoragetransfer()), cloudRetry);
+          addStep(
+              new DeleteStorageTransferServiceJobStep(flightBeanBag.getStoragetransfer()),
+              cloudRetry);
+          addStep(
+              new RemoveBucketRolesStep(flightBeanBag.getBucketCloneRolesService()), cloudRetry);
+        }
       }
     }
   }
